+++ src/index.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
-import App from "./components/App/App";
+import App from "./js/view/components/App/App";
 
 const root = ReactDOM.createRoot(
   document.getElementById("root") as HTMLElement
