+++ src/js/interfaces/Item.ts
