+++ src/js/interfaces/Store.ts
