+++ src/js/utils/fetchStorage.ts
