+++ src/js/utils/pure-helpers.ts
