+++ src/js/view/components/App/App.tsx
@@ -1,9 +1,9 @@
 import React from "react";
 import { Provider } from "react-redux";
-import store from "../../store/store";
+import store from "../../../stores/store";
 import Panel from "../Panel/Panel";
 
-import "../../styles/index.css";
+import "../../../../styles/index.css";
 
 function App() {
   return (
