+++ src/js/view/components/FindForm/FindForm.tsx
@@ -1,9 +1,9 @@
 import React, { useEffect, useState } from "react";
 import { useDispatch } from "react-redux";
-import { setList } from "../../store/store";
-import { List } from "../../interfaces";
-import { pullLocalStorage, pushLocalStorage } from "../../utils/localStorage";
-import { getStorageList } from "../../utils/storage";
+import { setList } from "../../../stores/store";
+import { List } from "../../../interfaces";
+import { pullLocalStorage, pushLocalStorage } from "../../../utils/localStorage";
+import { getStorageList } from "../../../utils/storage";
 
 const localStorageKey = "listKeyMemory";
