+++ src/js/view/components/index.ts
